backend/requirements.txt
azure-functions
azure-storage-blob>=12.19.0
azure-identity>=1.16.0
agent-framework[openai]>=1.0.0rc2
